Initialize warm liquid goo phase
I'm told this is actually circular reference but I'd have no idea of knowing that. Also, the workflow doesn't seem like a setup for circular reference. It just seems like change_scene_to_packed is breaking @export variables that point to PackedScenes in any instance where they are and make the project impossible to fix from the editor.